requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
django-qr-code==4.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Markdown==3.6
mysqlclient==2.2.4
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
segno==1.6.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
